I've created a test script to report on performance of certain constructs, to help debug and create test-cases for optimisation issues.
I have not found any documented coding standards for this project, so I have followed the standard I am used to.

Script includes usage instructions if called without arguments:
----------------------
> php test_performance.php
Tool to test DomPDF performance when rendering a large amount of HTML.  Uses repeated HTML snippets, normally representing common HTML elements, and reports on maximum memory usage and execution time for the various steps in the process.
The PDF is generated by repeating the HTML in the specified BODY_FILE the specified number of times.
    USAGE:   php test_performance.php BODY_FILE [ITERATIONS]
    EXAMPLE: php test_performance.php empty_div 100
Body files are placed in the 'html' sub-folder, which contains a number of useful examples.  The number of ITERATIONS may be omitted, in which case it defaults to 5000.
The following body files are currently available:
* break_tag
* empty
* empty_div
* empty_span
* whitespace
----------------------